+++ client/lib/platform/alipay_platform.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 class AlipayPlatform {
   static const MethodChannel _channel = MethodChannel('alipay');
 
-  static Future<bool> payV2(int price) async {
+  static Future<bool> payV2(double price) async {
     return await _channel.invokeMethod<bool>('payV2', price) ?? false;
   }
 }
